12.22

+++ yu-ai-agent/yu-ai-agent-master/src/main/java/com/yupi/yuaiagent/domin/entity/EsDocument.java
@@ -14,7 +14,7 @@ public class EsDocument {
     private String fileMd5;        // 文件指纹
     private Integer chunkId;       // 文本分块序号
     private String textContent;    // 文本内容
-    private float[] vector;        // 向量数据（768维）
+    private double[] vector;        // 向量数据（768维）
     private String modelVersion;   // 向量生成模型版本
     private String userId;         // 上传用户ID
     private String orgTag;         // 组织标签
@@ -30,7 +30,7 @@ public class EsDocument {
      * 完整构造函数，包含权限字段
      */
     public EsDocument(String id, String fileMd5, int chunkId, String content,
-                      float[] vector, String modelVersion,
+                      double[] vector, String modelVersion,
                       String userId, String orgTag, boolean isPublic) {
         this.id = id;
         this.fileMd5 = fileMd5;
@@ -42,6 +42,6 @@ public class EsDocument {
         this.orgTag = orgTag;
         this.isPublic = isPublic;
     }
-    
+
 
 }
